fix: resolve golangci-lint issues and split lint tasks in CI
- Fix 8 gosec G115: change errorCodeFromError to return int32, add
  safeIntToInt32 helper with bounds clamping for health metrics
- Fix staticcheck S1001: use copy() instead of manual loop
- Fix testifylint: use require.Empty / assert.NotEmpty
- Remove unused yKeys function from apilint
- Split Taskfile lint into lint:go and lint:api subtasks
- Add lint-api CI job running OpenAPI linter as separate pipeline step

+++ internal/api/handler_governance.go
@@ -215,7 +215,7 @@ func (h *APIHandler) DeleteLineageEdge(ctx context.Context, req DeleteLineageEdg
 func (h *APIHandler) PurgeLineage(ctx context.Context, req PurgeLineageRequestObject) (PurgeLineageResponseObject, error) {
 	deleted, err := h.lineage.PurgeOlderThan(ctx, int(req.Body.OlderThanDays))
 	if err != nil {
-		code := int32(errorCodeFromError(err))
+		code := errorCodeFromError(err)
 		return PurgeLineage403JSONResponse{Body: Error{Code: code, Message: err.Error()}, Headers: PurgeLineage403ResponseHeaders{XRateLimitLimit: defaultRateLimitLimit, XRateLimitRemaining: defaultRateLimitRemaining, XRateLimitReset: defaultRateLimitReset}}, nil
 	}
 	return PurgeLineage200JSONResponse{
